Only correct user to delete post

+++ app.py
@@ -100,23 +100,31 @@ def dashboard():
 
 
 @app.route('/posts/delete/<int:id>')
+@login_required
 def delete_post(id):
     post_to_delete = Posts.query.get_or_404(id)
+    id = current_user.id
+    if id == post_to_delete.poster.id:
+        try:
+            db.session.delete(post_to_delete)
+            db.session.commit()
 
-    try:
-        db.session.delete(post_to_delete)
-        db.session.commit()
+            # Return a message
+            flash("Blog Post Was Deleted!")
 
+            # Grab all the posts from the database
+            posts = Posts.query.order_by(Posts.date_posted)
+            return render_template("posts.html", posts=posts)
+        except:
+            # Return an error message
+            flash("There was a problem deleting post, try again...")
+
+            # Grab all the posts from the database
+            posts = Posts.query.order_by(Posts.date_posted)
+            return render_template("posts.html", posts=posts)
+    else:
         # Return a message
-        flash("Blog Post Was Deleted!")
-
-        # Grab all the posts from the database
-        posts = Posts.query.order_by(Posts.date_posted)
-        return render_template("posts.html", posts=posts)
-
-    except:
-        # Return an error message
-        flash("There was a problem deleting post, try again...")
+        flash("You Aren't Authorized To Delete That Post!")
 
         # Grab all the posts from the database
         posts = Posts.query.order_by(Posts.date_posted)
@@ -378,7 +386,6 @@ class Users(db.Model, UserMixin):
     password_hash = db.Column(db.String(128))
     # User Can Have Many Posts
     posts = db.relationship('Posts', backref='poster')
-
 
     @property
     def password(self):
